Semantic installed with bower, app.jsx and styles.scss import script and styles for semantic for build to app.js
Semantic installed with Bower does not compile from Less, this disallows themeing through the use of theme variables. Later on Semantic will be installed using NPM in /semantic directory, on build the css will be compiled and imported into styles.scss to be compiled with main gulp build command for app.